backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
loguru==0.7.3
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
email-validator==2.3.0
pymongo[srv]==4.6.1
bcrypt==4.1.2
itsdangerous==2.1.2
